minds: move _current_mngr_branch tests to fct_worktree_test
Follow-up to moving the FCT branch/vendor helpers into fct_worktree: relocate
the _current_mngr_branch unit tests to fct_worktree_test.py and refresh a stale
snapshot-script comment.
